requirements.txt
umap-learn
matplotlib
numpy
torch
Pillow